\docType{package}
\name{BigSEM-package}
\alias{BigSEM}
\alias{BigSEM-package}
\title{Constructing Large Systems of Structural Equations}
\description{This package implements the two-stage penalized least squares (2SPLS) method proposed by Chen, Zhang and Zhang (2016), which can construct a large system of structural equations, with the \eqn{k}-th structural equation as \eqn{Y_k = Y_{-k}\gamma_k+X_{S_k}\psi_{S_k}+\epsilon_k.}}
\references{
Chen, C., Zhang, M., and Zhang, D. (2016) \emph{A Two-Stage Penalized Least Squares Method for Constructing Large Systems of Structural Equations.} (Submitted)
}
\author{Chen Chen <chen1167@stat.purdue.edu>, Dabao Zhang <zhangdb@stat.purdue.edu>.}
